Apply Redux on CommentCreate
