+++ nyan-forum/src/views/CommentCreate/index.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react';
+import { connect } from 'react-redux';
 import styled from 'styled-components';
-import { db } from '../../firebase';
 
 import Button from '../../components/Button'
+import { createComment } from './actions';
 
 class CommentCreate extends Component {
   state = {
@@ -23,24 +24,9 @@ class CommentCreate extends Component {
     if (confirm) {
       const { postId } = this.props.match.params;
       const { author, content } = this.state;
-
-      // 새로운 댓글을 Comments Collection에 추가
-      const newCommentRef = db.collection('comments').doc();
-      await newCommentRef.set({
-        id: newCommentRef.id,
-        author,
-        content
-      });
-
-      // 새로 생성된 코멘트의 Id를 Post Collection의 comments에 추가
-      const postSnapshot = await db.collection('posts').doc(postId).get();
-      const commentsByPost = postSnapshot.data().comments;
-      await db.collection('posts').doc(postId).update({
-        comments: [
-          ...commentsByPost,
-          newCommentRef.id
-        ]
-      })
+      
+      const { dispatch } = this.props;
+      dispatch(createComment({ author, content, postId}));
 
       alert('성공적으로 등록되었습니다.');
       this.props.history.goBack();
@@ -123,4 +109,4 @@ const Textarea = styled.textarea`
   padding: 1rem;
 `;
 
-export default CommentCreate;
+export default connect()(CommentCreate);
